rename from robocup to soccer
