+++ module/purpose/Soccer/src/Soccer.hpp
@@ -1,5 +1,5 @@
-#ifndef MODULE_PURPOSE_ROBOCUP_HPP
-#define MODULE_PURPOSE_ROBOCUP_HPP
+#ifndef MODULE_PURPOSE_SOCCER_HPP
+#define MODULE_PURPOSE_SOCCER_HPP
 
 #include <nuclear>
 #include <string>
@@ -8,7 +8,7 @@
 
 namespace module::purpose {
 
-    class RoboCup : public ::extension::behaviour::BehaviourReactor {
+    class Soccer : public ::extension::behaviour::BehaviourReactor {
     private:
         /// @brief Smart enum for the robot's position
         struct Position {
@@ -43,10 +43,10 @@ namespace module::purpose {
         } cfg;
 
     public:
-        /// @brief Called by the powerplant to build and setup the RoboCup reactor.
-        explicit RoboCup(std::unique_ptr<NUClear::Environment> environment);
+        /// @brief Called by the powerplant to build and setup the Soccer reactor.
+        explicit Soccer(std::unique_ptr<NUClear::Environment> environment);
     };
 
 }  // namespace module::purpose
 
-#endif  // MODULE_PURPOSE_ROBOCUP_HPP
+#endif  // MODULE_PURPOSE_SOCCER_HPP
