+++ roles/webots/directormotion.role
@@ -42,5 +42,5 @@ nuclear_role(
   strategy::Ready
   # Purposes
   purpose::Striker
-  purpose::RoboCup
+  purpose::Soccer
 )
